Map CLI arguments to environment variables to make it easier to change one of them without changing all of them. I'm mostly thinking about the `docker run` use case of a standalone requestbin instance that uses in-memory storage--from a pre-built runscope/requestbin Docker Hub image that I hope exists soon!
This also allows people using the compose file to override selectively without overriding all of them. As with moving WEB_CONCURRENCY to the docker-compose file.

Also, I moved the MAX_REQUEST value of 2000 to the docker-compose.yml file and set a default of 0 in the Dockerfile--0 is the gunicorn default too which makes me think that'll be fewer surprises aside from a memory leak which as a container can be taken care of with limits instead of always restarting the worker even if there's not a problem. Again, this is for in-memory use cases. My rationale is that the impetus of requestbin is to inspect requests and for that people want to limit the loss of data. Yes, in-memory storage is inherently ephemeral but at least with these changes we limit the loss of data to the expected events: restarting the process (container) and a process (container) crash. And people can step up to redis (and a second container with a slightly more complicated setup) if they want to avoid that entirely.